+++ library/src/main/java/com/angcyo/library/unit/MmValueUnit.kt
@@ -15,7 +15,7 @@ import com.angcyo.library.isPlaceholderApplication
 class MmValueUnit : IValueUnit {
 
     /**获取刻度尺间隔的像素距离*/
-    override fun getGraduatedScaleGap(): Float = convertValueToPixel(1f)
+    override fun getGraduatedScaleGap(): Double = convertValueToPixel(1.0)
 
     /**将刻度索引值转换成多少个刻度单位*/
     override fun getGraduatedLabel(index: Int): String = "$index"
@@ -23,7 +23,7 @@ class MmValueUnit : IValueUnit {
     /**获取每个单位间隔刻度对应的像素大小
      * 将1个单位的值, 转换成屏幕像素点数值
      * [TypedValue.COMPLEX_UNIT_MM]*/
-    override fun convertValueToPixel(value: Float): Float {
+    override fun convertValueToPixel(value: Double): Double {
         val app = app()
         if (app.isPlaceholderApplication()) {
             return value
@@ -31,11 +31,11 @@ class MmValueUnit : IValueUnit {
         val dm: DisplayMetrics = app.resources.displayMetrics
         //1毫米等于多少像素                             //21.176456
         //1英寸等于多少像素, 1英寸=2.54厘米=25.4毫米      //537.882
-        return TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_MM, value, dm)
+        return TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_MM, value.toFloat(), dm).toDouble()
     }
 
     /**将value转换成对应单位的文本*/
-    override fun formattedValueUnit(value: Float): String {
+    override fun formattedValueUnit(value: Double): String {
         return when {
             //value.abs() / 100 > 1 -> "${(value / 100).decimal(2)}m"
             //value.abs() / 10 > 1 -> "${(value / 10).decimal(2)}cm"
